+++ src/components/points/DistanceComponent.jsx
@@ -8,7 +8,7 @@ function DistanceComponent({ id, latitude, longitude, classes, proximityToUnlock
   const { lat, long } = useContext(LatLongContext);
   const distance = useMemo(
     () => getDistanceBetweenCoordinates(lat, long, latitude, longitude),
-    [lat, long, latitude, longitude]
+    [lat, long, latitude, longitude],
   );
   function unlockThisPoint() {
     setListOfUnlocked([...listOfUnlocked, ...[id]]);